requirements.txt
Django==3.2.5
django-environ==0.4.5
psycopg2-binary==2.9.1
djangorestframework==3.12.4
django-cors-headers==3.7.0